feature: small number censoring
censored small numbers/rates for the CTTown_Alert.csv for ODP.  I NA'd out the values if they are RateCategory 1.  Need to confirm how datawrapper would handle NA's with Gary.

I also created a little thursday specific setup section that combines elr with the geocoding and moved that to the top so if anyone  runs the top of this they can run each piece individually and not have to fish around